suggest ViewFinder - head for headless browsers
This is probably rude to suggest as I'm the maker of ViewFinder, and I'm sure you think it's irrelevant here. Just saw your list on HN and thought this might be a useful addition for people looking to interact with headless browsers as if they were a normal *headful* browser.